- update webpack config - move Sandbox to VuePress folder
