+++ docs/sandbox.md
@@ -1,0 +1,6 @@
+---
+sidebar: false
+editLink: false
+layout: Sandbox
+---
+
